[ui] Display gannt chart of queued runs, move queued banner to logs panel (#28831)
related:
https://linear.app/dagster-labs/issue/OPER-1560/display-gantt-chart-for-queued-runs

## Summary & Motivation

The execution plan data is available for a queued run, so we may as well
show the gantt chart. The logs on the other hand seem to always be
empty, so this PR moves the “queued” empty state there.

## How I Tested These Changes

Tested with queued runs manually since it's a small visual change.

## Changelog

[ui] The Dagster Run UI now allows you to view the execution plan of a
queued run

(Plan with one step is visible below, and the empty state is at the
bottom)

![image](https://github.com/user-attachments/assets/93991030-5b61-45f8-9a7a-b52da6b06762)

Co-authored-by: bengotow <bgotow@elementl.com>